move pages around for launch
