+++ pages/app/projects/applications/section/[id].tsx
@@ -1,4 +1,4 @@
-import SideBar from '@components/common/Home/SideBar';
+import SideBar from '@components/common/Feed/SideBar';
 import DetailContent from '@components/common/Project/MyApplication/DetailContent';
 
 const Section = () => {
